rename constants to status_type
